# lib/api-zod/src/generated/types/downloadJobResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DownloadJobResponseStatus } from "./downloadJobResponseStatus";

export interface DownloadJobResponse {
  jobId: string;
  status: DownloadJobResponseStatus;
  progress?: number;
  filename?: string;
  filesize?: number;
  error?: string;
  title?: string;
  mode?: string;
}