# lib/api-zod/src/generated/types/downloadRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DownloadRequestAudioQuality } from "./downloadRequestAudioQuality";
import type { DownloadRequestMode } from "./downloadRequestMode";
import type { DownloadRequestVideoQuality } from "./downloadRequestVideoQuality";

export interface DownloadRequest {
  url: string;
  mode: DownloadRequestMode;
  videoQuality?: DownloadRequestVideoQuality;
  audioQuality?: DownloadRequestAudioQuality;
}